added sway and alacritty configs. various modifications

+++ .local/installers/void-install.sh
@@ -8,25 +8,24 @@ sudo xbps-install -S wget curl
 #enable repos
 #clear
 echo "enable nonfree repo?(y/N)"
-read input
-if [ $input == 'y' ] || [ $input == 'Y' ]; then
+read NONFREE 
+if [ "$NONFREE" = "y" ]; then
     sudo xbps-install -Sy void-repo-nonfree
-	sudo xbps-install -Sy unrar
 fi
 
-#clear
-#echo "enable multilib repo?(y/N)"
-#read input
-#if [ $input == 'y' ] || [ $input == 'Y' ]; then
-#    sudo xbps-install -Sy void-repo-multilib   
-#fi
+clear
+echo "enable multilib repo?(y/N)"
+read MULTILIB 
+if [ "$MULTILIB" = "y" ]; then
+    sudo xbps-install -Sy void-repo-multilib   
+fi
 
-#clear
-#echo "enable nonfree multilib repo?(y/N)"
-#read input
-#if [ $input == 'y' ] || [ $input == 'Y' ]; then
-#    sudo xbps-install -Sy void-repo-multilib-nonfree
-#fi
+clear
+echo "enable nonfree multilib repo?(y/N)"
+read NFMULTILIB 
+if [ "$NFMULTILIB" == "y" ]; then
+    sudo xbps-install -Sy void-repo-multilib-nonfree
+fi
 
 
 #make directories
@@ -46,15 +45,17 @@ if [ $input == 'y' ] || [ $input == 'Y' ]; then
 #sudo xbps-install -Sy xorg xterm
 sudo xbps-install -Sy xorg-minimal xrdb xsetroot xterm xprop 
 
-#install devel tools
-sudo xbps-install -Sy base-devel make libXft-devel libX11-devel libXinerama-devel patch
-
 #sound programs
-sudo xbps-install -Sy alsa-utils apulse #pamixer pulsemixer  
+sudo xbps-install -Sy alsa-utils apulse 
 
 #install system tools
-sudo xbps-install -Sy feh udevil picom redshift ffmpegthumbnailer ConsoleKit2 apparmor sxhkd dunst scron rsync
-sudo xbps-install -Sy gst-libav xdg-utils GConf
+sudo xbps-install -Sy freetype ConsoleKit2 apparmor gst-libav xdg-utils GConf wget curl chrony
+
+["$NONFREE" = "y" ] && sudo xbps-install -Sy unrar
+
+
+sudo xbps-install -Sy feh udevil picom redshift ffmpegthumbnailer sxhkd dunst scron rsync
+
 
 # appimage needed this
 sudo xbps-install -Sy dbus-glib #nss startup-notification
@@ -73,9 +74,6 @@ if [ $input == 'y' ] || [ $input == 'Y' ]; then
 	sudo xbps-install -Sy tlp
     sudo ln -s /etc/sv/tlp /var/service
 fi
-
-#install NetworkManager
-#sudo xbps-install -S NetworkManager
 
 clear
 echo "install and setup ufw firewall? (y/N)"
@@ -104,6 +102,8 @@ sudo ufw allow 56881:56889/tcp
 sudo ufw allow 27000:27036/udp
 sudo ufw allow 27036:27037/tcp
 sudo ufw allow 4380/udp
+
+sudo ufw enable
 fi
 
 fi
@@ -117,7 +117,7 @@ if [ $nvidiain == 'y' ] || [ $nvidiain == 'Y' ]; then
 
     sudo xbps-install -Sy void-repo-nonfree void-repo-multilib void-repo-multilib-nonfree 
 	sudo xbps-install -Sy xf86-video-nouveau nvidia nvidia-libs-32bit
-#sudo xbps-install vulkan-loader vulkan-loader-32bit
+    sudo xbps-install vulkan-loader vulkan-loader-32bit
 fi
 
 
@@ -136,7 +136,9 @@ if [ $amdin == 'y' ] || [ $amdin == 'Y' ]; then
 
 fi
 
-sudo xbps-install -Sy vulkan-loader vulkan-loader-32bit
+sudo xbps-install -Sy vulkan-loader
+
+[ "$MULTILIB" = "y" ] && sudo xbps-install -Sy vulkan-loader-32bit
 
 clear
 echo "install flatpak and setup flathub?(y/N)"
@@ -157,19 +159,6 @@ if [ $input == 'y' ] || [ $input == 'Y' ]; then
 	sudo ln -s /etc/sv/popcorn /var/service/
 fi
 
-
-### install themes ###
-
-#sudo xbps-install -Sy adwaita-icon-theme Adapta gtk-engine-murrine
-
-#install qt for apps that need it
-#sudo xbps-install -Sy qt5
-
-#install fonts
-sudo xbps-install -Sy freetype
-#sudo xbps-install -S noto-fonts-cjk noto-fonts-emoji noto-fonts-ttf noto-fonts-ttf-extra
-#sudo xbps-install -Sy font-libertine-otf font-inconsolata-otf font-symbola font-hack-ttf nerd-fonts-otf 
-#sudo xbps-install -Sy ipafont-fonts-otf
 
 #install fish shell
 clear
@@ -207,41 +196,26 @@ read input
 if [ $input == 'y' ] || [ $input == 'Y' ]; then
 
 #cli programs to isntall
-sudo xbps-install -Sy vim htop curl wget tealdeer ufetch
+sudo xbps-install -Sy vim htop tealdeer ufetch
 sudo xbps-install -Sy mpv cmus lf mtm #pulsemixer pamixer mtm
 sudo xbps-install -Sy ImageMagick jp2a ffmpegthumbnailer #for ascii image previews and video thumbnails in lf
 sudo xbps-install -Sy zathura zathura-cb zathura-pdf-mupdf
-sudo xbps-install -Sy youtube-viewer youtube-dl
+sudo xbps-install -Sy youtube-dl
 sudo xbps-install -Sy transmission transmission-remote-cli
 #sudo xbps-install -Sy trackma
 
-#sudo xbps-install -Sy termdown spt
-#install vim plug for neovim
-#curl -fLo ~/.local/share/nvim/site/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
 fi
 
 
 clear
-echo "install my gui programs?(y/N)"
+echo "install gaming programs(wine lutris steam)? (y/N)"
+echo "this will enable nonfree repos"
 read input
 if [ $input == 'y' ] || [ $input == 'Y' ]; then
-#gui programs to install
-sudo xbps-install -Sy lxappearance #firefox #gnome-disk-utility  
+    sudo xbps-install -Sy void-repo-nonfree void-repo-multilib void-repo-multilib-nonfree 
+	sudo xbps-install -Sy wine-32bit wine-gecko wone-mono winetricks protontricks 
+	sudo xbps-install -Sy lutris steam
 fi
-
-#gaming
-#echo "install wine, steam and lutris for gaming? (y/N)"
-#read input
-#if [ $input == 'y' ] || [ $input == 'Y' ]; then
-#sudo xbps-install -Sy libpulseaudio-32bit libtxc_dxtn-32bit fontconfig-32bit libavcodec-32bit libavformat-32bit libavresample-32bit libavutil-32bit 
-#sudo xbps-install -Sy lutris
-#sudo xbps-install -Sy wine-32bit libwine-32bit wine-gecko wine-mono winetricks protontricks
-#sudo xbps-install mono #stardew valley (smapi)
-#fi
-
-
-#development
-#sudo xbps-install -Sy ncurses-devel mdBook tcc
 
 #Haskell
 #sudo xbps-install -Sy ghc cabal-install stack
@@ -250,8 +224,12 @@ fi
 #sudo xbps-install -Sy lua love luarocks
 
 #muttwizard
-sudo xbps-install -Sy neomutt isync msmtp pass urlview
-
+clear
+echo "install muttwizard dependencies? (y/N)"
+read input
+if [ $input == 'y' ] || [ $input == 'Y' ]; then
+	sudo xbps-install -Sy neomutt isync msmtp pass urlview
+fi
 
 ### SYSTEM SETUP ### 
 
@@ -259,6 +237,7 @@ sudo xbps-install -Sy neomutt isync msmtp pass urlview
 sudo ln -s /etc/sv/dbus /var/service
 sudo ln -s /etc/sv/cgmanager /var/service/
 sudo ln -s /etc/sv/consolekit /var/service/
+sudo ln -s /etc/sv/chronyd /var/service/
 
 ### set limits.conf for esync
 sudo sh -c "echo '$USER hard nofile 524288' >> /etc/security/limits.conf"
@@ -286,11 +265,15 @@ if [ $input == 'y' ] || [ $input == 'Y' ]; then
 	sudo ln -s ~/.config/.sysconf/etc/X11/xorg.conf.d/50-mouse-acceleration.conf /etc/X11/xorg.conf.d/
 fi
 
-#suckless dependancies
+###suckless-install.sh dependancies
 sudo xbps-install -S base-devel make libXft-devel libX11-devel libXinerama-devel patch
 
-#surf dependancies
+  #surf dependencies
 sudo xbps-install -S gtk+-devel webkit2gtk-devel gcr-devel
+
+  #straw-viewer dependencies
+sudo xbps-install -S perl-Module-Build
+
 
 
 sh ~/.local/installers/suckless-install.sh
